Register types of audio into the type registry
